+++ backend/MovieBooking/src/main/java/com/harshit/moviebooking/entity/User.java
@@ -57,4 +57,10 @@ public class User {
 
     @OneToMany(mappedBy = "user")
     private List<Rating> ratings;
+
+    @OneToMany(mappedBy = "user")
+    private List<Review> reviews;
+
+    @OneToMany(mappedBy = "user")
+    private List<WatchlistEntry> watchlistEntries;
 }